fix i32_load* unittest
* 修正 i32_load* unittest 的檢查範圍
* 修正 i32_load16_s 的 typo
* 修正 i32_load16_s 的錯誤的型別轉換

+++ src/lib/core/runtime/i32_load16_s.c
@@ -16,7 +16,7 @@ int runtime_i32_load16_s(Stack* stack, MemInst* memory, uint32_t offset, uint32_
         free(value1);
         return -1;
     }
-    int8_t* data = (int8_t*)(memory->data->data + ea);
+    int16_t* data = (int16_t*)(memory->data->data + ea);
     stack->entries->push(stack->entries, new_i32Value(*data));
     free(value1);
     return 0;
